+++ src/source/TMC2130Stepper.cpp
@@ -1,4 +1,4 @@
-#include "TMCStepper.h"
+#include "TMCStepper_main.h"
 #include "TMC_MACROS.h"
 
 uint32_t TMC2130Stepper::spi_speed = 16000000/8;
